move: generate domain dir, view dir
